% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancetest_function.R
\name{balance_fn}
\alias{balance_fn}
\title{Balance Test Table Production}
\usage{
balance_fn(dataset, assignment)
}
\arguments{
\item{dataset}{is your dataframe}

\item{assignment}{is the name of the variable used to identify treatment and control.}
}
\description{
This function takes in a dataset and treatment assignment variable, and produces balance tests and control/treatment means in table form.
}
\examples{
balance_fn()
}
