% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_helper_functions.R
\name{causal_forest_plots}
\alias{causal_forest_plots}
\title{Causal Forest Plotting}
\usage{
causal_forest_plots(df, y, x_var, exclude = NULL, split_r = 0.9,
  seed_1 = 24010, seed_2 = 39028, seed_3 = 45829)
}
\arguments{
\item{df}{is your dataframe}

\item{y}{is the outcome being predicted.}

\item{x_var}{is the specific feature to plot against that outcome}

\item{exclude}{takes any variables you want to omit from the sample.}

\item{split_r}{is the training/test proportion split you want. Defaults to .9.}

\item{seed_1}{and those that follow allow you to see the random partition for reproducibility.}
}
\description{
This function takes in a dataframe with features and an outcome. It produces predictions for a test set, and plots them against the features.
}
\examples{
causal_forest_plots()
}
