% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest_function.R
\name{ftest_fn}
\alias{ftest_fn}
\title{F Test Function}
\usage{
ftest_fn(sourcefile, varlist, assignvar)
}
\arguments{
\item{sourcefile}{is your dataframe, should include a variable for treatment assignment and whatever demographics.}

\item{varlist}{is a vector of your covariates (things you are controlling for)- all should be the names of columns in your dataset.}

\item{assignvar}{is the specific column that specifies the treatment/control split in your sample. Also, your blocking variable should be "blocknum".}
}
\description{
This function computes the f-test p-value for baseline characteristics between treatment and control.
}
\examples{
ftest_fn(dataset, c("var1", "var2", "var3"), "dbam")
}
