% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kling_as_function.R
\name{itt_for_kling_p3}
\alias{itt_for_kling_p3}
\title{Kling Model with Three Participation Variables}
\usage{
itt_for_kling_p3(outcome, participation1, participation2, participation3,
  grouping, sizefile, weight, number_to_crop, sourcefile, weight_final,
  omit_type)
}
\arguments{
\item{number_to_crop}{should equal the number of grouping categories you have. It's for selecting the interactions only out of the model.}
}
\description{
This function is generally the same as itt_for_kling but it allows for three years and three participation vars.
}
\examples{
School Level MTO
itt_for_kling_p3(outcome = "mathxil_z_post2", participation1 = "treat_post1_single", participation2 = "treat_post2",
participation3 = "treat_post3", grouping = "schlid", sizefile = schoolsizes, weight = "mtoweight_calculation", number_to_crop = 12,
sourcefile = master_dataset_1_math15_no2531, weight_final = "stderr_scores_weight", "omits")
}
