% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kling_as_function.R
\name{itt_for_kling_p4}
\alias{itt_for_kling_p4}
\title{Kling Model with Two Grouping Variables}
\usage{
itt_for_kling_p4(outcome, participation1, participation2, grouping1, grouping2,
  sizefile, weight, number_to_crop, sourcefile, weight_final, omit_type)
}
\arguments{
\item{number_to_crop}{should equal the number of grouping categories you have. It's for selecting the interactions only out of the model.}
}
\description{
This function is generally the same as itt_for_kling but it allows for mixing block and school (or whatever) grouping variables.
}
\examples{
School Level
itt_for_kling_p4(outcome = "mathxil_z_post2", participation1 = "treat_post1", participation2 = "treat_post2",
grouping1 = "schlid", grouping2 = "blocknum", sizefile = schoolsizes, weight = "precision", number_to_crop = 42,
sourcefile = master_dataset_1_math15_allblocks, weight_final = "stderr_scores_weight", "noomits_nome")
}
