% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kling_as_function.R
\name{itt_for_kling_p6}
\alias{itt_for_kling_p6}
\title{Kling Model with Two Grouping Variables, Instrumenting Var 2}
\usage{
itt_for_kling_p6(outcome, participation1, participation2, grouping1, grouping2,
  sizefile, weight, number_to_crop, sourcefile, weight_final, omit_type)
}
\arguments{
\item{number_to_crop}{should equal the number of grouping categories you have. It's for selecting the interactions only out of the model.}
}
\description{
This function is generally the same as itt_for_kling_p4 but it puts grouping variable 2 in the ITT models and instruments for it in the ivreg.
This variation seems to have produced good results for the Match/BAM project.
}
\examples{
Block Level
itt_for_kling_p6(outcome = "mathxil_z_post2", participation1 = "treat_post1", participation2 = "treat_post2",
grouping1 = "schlid", grouping2 = "blocknum", sizefile = schoolsizes, weight = "precision", number_to_crop = 42,
sourcefile = master_dataset_1_math15_allblocks, weight_final = "stderr_scores_weight", "noomits_nome")
}
