% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_totanditt.R
\name{itt_tot_model}
\alias{itt_tot_model}
\title{ITT and TOT Model Function}
\usage{
itt_tot_model(sourcefile, I, varlist, assignvar, treatvar, weight = NULL)
}
\arguments{
\item{sourcefile}{is your dataframe, should include a variable for treatment assignment and whatever demographics.}

\item{I}{is the outcome being predicted.}

\item{varlist}{is a vector of your covariates (things you are controlling for)- all should be the names of columns in your dataset.}

\item{assignvar}{is the specific column that specifies the treatment/control split in your sample. Also, your blocking variable should be "blocknum".}

\item{treatvar}{is the column specifying actual treatment dosage regardless of assignment.}
}
\description{
This function computes multiple ITT and TOT models at once. Produces robust standard errors.
}
\examples{
itt_tot_model()
}
