% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_style.R
\name{model_table_style}
\alias{model_table_style}
\title{Produce ITT and TOT tables in Crime Lab style.}
\usage{
model_table_style(tablename, colpositions, titletext = "Table",
  footnotetext = "")
}
\description{
Produce ITT and TOT tables in Crime Lab style.
}
