% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationtest.R
\name{permtest_multiple}
\alias{permtest_multiple}
\title{Permutation Test Function}
\usage{
permtest_multiple(data, indices, M)
}
\arguments{
\item{data}{is your dataframe.}

\item{M}{is the outcome being tested.}
}
\description{
This function computes permutation tests, using sample with replacement, keeping the blocks balanced. You need to edit the covariates inside the function for use in other projects.
}
\examples{
results_rescaled_math_post1  <- boot(data=master_dataset_1, statistic=bs_post, strata=master_dataset_1$blocknum,
R=100000, M="d$rescaled_math_post1", sim="permutation")
Result: results_rescaled_math_post1$t
}
