% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest_function.R
\name{ftest_fn}
\alias{ftest_fn}
\title{F Test Function used in BAM and Match}
\usage{
ftest_fn(sourcefile)
}
\arguments{
\item{sourcefile}{is your dataframe, should include a variable for treatment assignment and whatever demographics.}
}
\description{
This function computes the f-test p-value for baseline characteristics between treatment and control.
}
\examples{
ftest_fn()
}
